requirements.txt
# Speech-to-Text
vosk
openai-whisper
torch
torchaudio
torchvision

# Text Analysis/NLP
spacy
transformers
spacy-transformers

# Text-to-Speech
TTS

# API Development
fastapi
uvicorn
flask
graphene

# Database
django 
djangorestframework

# General Utilities
numpy
pandas
pydub
ffmpeg-python
# base64